intel: Fix broken acceleration if a NVIDIA GPU is present
If the Intel DDX gets started while NVIDIA's driver is loaded, the NVIDIA driver will
take control of the DRM subsystem and set itself as "card0". This is bad as the DDX isn't adapted for such change.

Luckily the driver seems to leave render nodes alone.
Add an extra check and logic to detect their driver and recalculate the render node path.