intel/driver: Optimize `intel_device_info`
Out of all the things that could break the structification, it was _this_ guy.

Cool, instead force alignment for the struct and shave off some additional bits that we won't need.
`0xff` is now a special value for `gen` as it will be likely used by EGA to flag modern Intel graphics.