sna/present: Try to actually fix the modifier mismatch
Turns out that whatever demented thing DXVK is doing to trigger this in Xorg, it made things bad.

Instead, revert to the old logic during flip() but make the check no longer depend on an asynchronous flip,
which is no longer the case for some reason... Instead make the check a simple, are we linear, is the user
not requesting a linear FB and is the pixmap from the hellscape known as DRI3.

I wish the Linux desktop stack didn't suck so much.

p.s. We might need to also include the check in check_flip2() but that's for future work.

Fixes: 027416ab3551fac1b7e22c7e253c8134c2f222b7