sna: Actually actually actually fix cff2b16267597ccaeed6be6b8035f001f08f7a90
Turns out that Intel hardware is a trainwreck (holding self back here) and trying to do PRIME sharing
is a mess considering that we have so much going on in SNA land.

Here are the key points to note:
- We cannot do asynchronous scanout of LINEAR modifiers. (hardware limitation)
- BLT (blitter) is a thorn in the back, since it can introduce latency.
- The previous workaround in `sna_present_check_flip2` didn't work since the code path would almost never get triggered.
- Doing it in `get_flip_bo` also didn't really help either.

Since both of these are a mess and I've looked at the source code while being slightly more awake than usual,
I can say that the cause of all our problems is DRI3, since it's pixmap sharing that is causing it.

I could in theory try to enable the shadow flipping logic and do another `slave_pixmap` logic to avoid the issue but this would
require me to suffer more.

For now, I've made new internal Pixmap usage flags:
- `SNA_PIXMAP_USAGE_DRI3_IMPORT`
- `SNA_PIXMAP_USAGE_DRI3_IMPORT2`

`SNA_PIXMAP_USAGE_DRI3_IMPORT` is set when a pixmap is imported via `sna_dri3_pixmap_from_fd()`, since this is the origin
of our problems.

When the pixmap is flagged the following happens:
- Scanout of the pixmap is no longer possible asynchronously
- Copies via `sna_copy_boxes` will no longer use the blitter engine.

The second flag is for future use and currently has zero purpose.

With that said, it seems that we are still limited (on NVIDIA only?) to our main display framerate with this patch,
however motion is WAY WAY WAY more responsive as we are no longer dealing with the thorn.

I think this should be the end of the saga until I finish up DRI3 v1.2 support.

tl;dr pain

(plus a struct cleanup since I missed one or two, some debugging work to make my life easier in the future and a fix for an assert in the AVX2 blitter path)