various: Disable i810 support by default.
This is ancient code for ancient hardware that no-one should be running or be able to run as DRI1 is dead and UMS is dead as of Linux 6.8.

We'll keep the code and the option to build it for historical purposes but I don't want any of this by default.